Create a input Component
