+++ 316-es6-spread-operator-practice/src/components/App.jsx
@@ -1,45 +1,26 @@
 import React, { useState } from "react";
-import RemoveButton from "./RemoveButton";
 import ToDoItem from "./ToDoItem";
+import InputArea from "./InputArea";
 
 function App() {
-  const [toDoItem, setToDoItem] = useState("");
   const [toDoList, setToDoList] = useState(["Buy eggs"]);
-  function handleChange(event) {
-    setToDoItem(event.target.value);
-  }
 
-  function handleClick() {
-    setToDoList((prevList) => [...prevList, toDoItem]);
-    console.log(toDoItem);
-    setToDoItem("");
-  }
-  function handleKeyDown(event) {
-    event.key === "Enter" && handleClick();
-  }
+  function addItem(newItem) {
+    setToDoList((prevList) => [...prevList, newItem]);
+    console.log(newItem);
+  };
 
   function removeItem(index) {
     console.log(index);
     setToDoList((prevList) => prevList.filter((_, i) => i!== index));
-  }
+  };
 
   return (
     <div className="container">
       <div className="heading">
         <h1>To-Do List</h1>
       </div>
-      <div className="form">
-        <input
-          type="text"
-          placeholder="Do something"
-          onChange={handleChange}
-          value={toDoItem}
-          onKeyDown={handleKeyDown}
-        />
-        <button onClick={handleClick}>
-          <span>Add</span>
-        </button>
-      </div>
+      <InputArea addItem={addItem} />
       <div>
         <ul>
           {toDoList.map((item, index) => (
